Moving common test code to test utils from text downloaders
